# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /push-token:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register or update an Expo push token for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPushTokenRequest"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/streak-check:
    post:
      operationId: checkStreakNudge
      tags: [notifications]
      summary: Check if a streak nudge should be sent for the user today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreakCheckRequest"
      responses:
        "200":
          description: Streak nudge check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakCheckResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/upload-url:
    post:
      operationId: getUploadUrl
      tags: [storage]
      summary: |
        Get a presigned GCS PUT URL for direct meal photo upload.
        Step 1 of 2: call this to obtain uploadUrl and objectKey.
        Step 2 of 2: PUT the image bytes directly to uploadUrl (Content-Type: image/jpeg).
        Photo upload is independent of food analysis — run both in parallel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/objects/{objectKey}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve a stored meal photo by its object key
      parameters:
        - name: objectKey
          in: path
          required: true
          schema:
            type: string
          description: The objectKey returned by getUploadUrl (e.g. meal-photos/userId/uuid.jpg)
      responses:
        "200":
          description: Image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    RegisterPushTokenRequest:
      type: object
      properties:
        userId:
          type: string
        token:
          type: string
      required:
        - userId
        - token
    StreakCheckRequest:
      type: object
      properties:
        userId:
          type: string
        localDate:
          type: string
          description: Current local date in YYYY-MM-DD format
      required:
        - userId
        - localDate
    StreakCheckResponse:
      type: object
      properties:
        shouldNudge:
          type: boolean
        streakDays:
          type: integer
      required:
        - shouldNudge
        - streakDays
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        gender:
          type: string
        birthday:
          type: string
        heightCm:
          type: number
        weightKg:
          type: number
        targetWeightKg:
          type: number
        goal:
          type: string
        activityLevel:
          type: string
        dailyCalorieTarget:
          type: integer
        dailyProteinTarget:
          type: number
        dailyCarbsTarget:
          type: number
        dailyFatTarget:
          type: number
      required:
        - userId
        - gender
        - birthday
        - heightCm
        - weightKg
        - targetWeightKg
        - goal
        - activityLevel
        - dailyCalorieTarget
        - dailyProteinTarget
        - dailyCarbsTarget
        - dailyFatTarget
    UpdateProfileRequest:
      type: object
      properties:
        userId:
          type: string
        gender:
          type: string
        birthday:
          type: string
        heightCm:
          type: number
        weightKg:
          type: number
        targetWeightKg:
          type: number
        goal:
          type: string
        activityLevel:
          type: string
      required:
        - userId
        - gender
        - birthday
        - heightCm
        - weightKg
        - targetWeightKg
        - goal
        - activityLevel
    UploadUrlRequest:
      type: object
      description: Request a presigned GCS PUT URL for direct photo upload
      properties:
        userId:
          type: string
          description: User UUID v4 (key is scoped to this user)
        fileName:
          type: string
          description: Original file name (used for logging only)
        contentType:
          type: string
          description: MIME type of the image (e.g. image/jpeg)
      required:
        - userId
        - fileName
        - contentType
    UploadUrlResponse:
      type: object
      description: |
        Presigned GCS upload URL. PUT image bytes directly to uploadUrl
        with Content-Type matching the requested contentType.
      properties:
        uploadUrl:
          type: string
          description: Presigned GCS PUT URL — upload image bytes directly here (15 min TTL)
        objectKey:
          type: string
          description: Storage key for the uploaded object (save this for the meal entry)
        publicUrl:
          type: string
          description: API serving URL for the photo after upload completes
      required:
        - uploadUrl
        - objectKey
        - publicUrl